Rename topk to topkTokens
